+++ cpu-app/Models/expense-report/DynamicsScheduleGLineItemCollectionPost.cs
@@ -7,10 +7,10 @@ namespace Gov.Cscp.Victims.Public.Models
 {
 	public class DynamicsScheduleGLineItemCollectionPost
 	{
-		public float vsd_actualexpensescurrentquarter { get; set; }
-		public float vsd_quarterlyvariance { get; set; }
-		public float vsd_actualexpendituresyeartodate { get; set; }
-		public float vsd_yeartodatevariance { get; set; }
+		public decimal vsd_actualexpensescurrentquarter { get; set; }
+		public decimal vsd_quarterlyvariance { get; set; }
+		public decimal vsd_actualexpendituresyeartodate { get; set; }
+		public decimal vsd_yeartodatevariance { get; set; }
 		public string fortunecookietype { get { return "Microsoft.Dynamics.CRM.vsd_scheduleglineitem"; } }
 		public string vsd_scheduleglineitemid { get; set; }
 		public string vsd_explanationforvariance { get; set; }
